Partial conversion to multiple stimulus lists
Converted the experiments to use a list of stimulus lists by default. A specific variable ?expnum=<n>, n counting from 1, is used to select the actual list to be used (defaults to 1). The script to generate stimulus list .js files will generate the list based on a column ExpNum. The ability to enter a different name will be implemented later. If no such column is present/given, the script will output only a single list.
This version has only partial functionality in generating lists of stimulus lists.